test(input): move tmux applier tests beside their private appliers
The three tmux appliers (apply_tmux_shortcuts, apply_tmux_copy_mode,
apply_tmux_forward) are now private, registered by ShortcutsTmuxModePlugin
in shortcuts/tmux.rs. Their tests still lived in tmux/input.rs and imported
them across the module boundary, which no longer compiles (E0603).

Move the applier tests and their helpers (TmuxCaptured, tmux_dispatch_app,
dispatch) into a #[cfg(test)] mod tests in shortcuts/tmux.rs, where super::*
reaches the private functions. The wheel/aggregate/consume and
direction-mapping tests stay in input.rs.
